Add Period to StoxLandingData

Add a column to StoxLandingData with a temporal category, such as quarters.

Usage

AddPeriodStoxLanding(
  StoxLandingData,
  TemporalDefinition,
  ColumnName = character()
)

Arguments

StoxLandingData

StoxLandingData data which will be annotated.

TemporalDefinition

TemporalDefinition definition of temporal category.

ColumnName

Name of the added column. Defaults to Period.

Value

StoxLandingData with column appended. See StoxLandingData.

Details

A column with a temporal category will be added based on the column 'CatchDate' in 'StoxLandingData'. Temporal definitions (TemporalDefinition) may be produced by DefinePeriod

Since it may be useful to use different temporal categories for instance for parameterization and reporting two choices are offered for the name of the added column (see argument 'ColumnName').

See also