Analytical Population Estimate Data

Analytical estimates for a population

SampleSummary

Stratum

Stratum that summary is provided for.

PSUDomain

PSU domain that summary is provided for.

Samples

The number of PSUs sampled in the domain.

PSUDomainSize

The estimated number of PSUs in the domain.

PSURelativeDomainSize

The estimated fraction of PSUs in the domain, relative to the number of PSUs in the stratum.

Abundance

Stratum

Stratum that abundance is provided for.

Domain

The domain that the abundance is provided for.

Abundance

The estimated number of individuals in the domain and stratum.

Frequency

The estimated fraction of individuals in the domain, relative ot the total number in stratum.

Variables

Stratum

The stratum that estimates are provided for.

Domain

The domain that estimates are provided for.

Variable

The variable (measurment) that estimates are provided for.

Total

The estimated total value of the variable in domain and stratum.

Mean

The estimated mean value fo the variable in domain and stratum.

AbundanceCovariance

Stratum

The stratum that covariances are provided for.

Domain1

A domain that covariances are provided for.

Domain2

A domain that covariances are provided for.

AbundanceCovariance

The estimated covariance of abundance between Domain1 and Domain2.

FrequencyCovariance

The estimated covariance of frequency between Domain1 and Domain2.

VariableCovariance

Stratum

The stratum that covariances are provied for.

Domain1

A domain that covariances are provided for.

Domain2

A domain that covariances are provided for.

Variable1

A variable that covariances are provided for.

Variable2

A variable that covariances are provided for.

TotalCovariance

The estimated covariance of total value of Variable1 in Domain1 and Variable2 in Domain2

MeanCovariance

The estimated covariance of the mean value of Variable1 in Domain1 and Variable2 in Domain2

StratificationVariables

Stratum

A stratum, as identified in other tables.

StratificationVariables

Columns that relate the stratum to data records.

DomainVariables

Domain

A domain, as identified in other tables.

DomainVariables

Columns that relate the domain to data records.