LOCODE table (LocodeTable)

Table of United Nations Codes for Trade and Transport Locations (UN/LOCODE)

Country

ISO 3166 alpha-2 Country Code for the country in which the location is placed.

Location

3-character code identifying the location within country

NameWoDiacritics

Name of location without diacritic signs.

Subdivision

ISO 1-3 character alphabetic and/or numeric code for the administrative division of the country concerned (the latter part of the complete ISO 3166-2 code element (after the hyphen).)

Status

code for status of the code. E.g. Codes starting with A are various official approvals.

Function

string with codes classifying the function of the location. If this contains 1, the location has a seaport.

Coordinates

latitude and longitude in degrees and minutes.

Remarks

general remark about location or change in coding

latitude

latitude in decimal degrees

longitude

longitude in decimal degrees