Summary statistics for simulated parameters (ParameterizationSummaryData)

Summary statistics for (potentially multi-chained) simulated parameters, such as MCMC simulations with Reca. 'chains' in this respect refers to statistically independent simulations.

list with two members 'ParameterSummary' and 'RunParameters', which both all data.tables. 'ParameterSummary' contains parameter statistics for simulations and have the following columns:

Parameter

Identifies the parameter that is summarized.

Mean

Mean of the parameter

Variance

Variance of the parameter

chainId

Identifies the parameterization chain.

RunParameters summarizes global control parameters for each chain and has columns:

chainId

Identifies the parameterization chain.

Iterations

The number of iterations for parameter simulation

See also