Fisheries dependent Catch At Age Report (ReportFdaCatchAtAgeData)

A list with two members: 'NbyAge' and 'GroupingVariables'.

NbyAge

A ReportFdaData table with reported <Statistic> being 'CatchAtAge': the total catch at age in numbers.

GroupingVariables

Any specified Grouping variables.

Units are configurable, and can be inspected by ~getUnit

'GroupingVariables' is a data.table with a column containing the names of any aggregation variables.