Fisheries dependent analysis report (ReportFdaData)

Results from catch estimations. The results may be presented decomposed on combinations of aggregation variables, such as gear, area, stock etc.

list with two members 'FdaReport' and 'GroupingVariables'. 'FdaReport' is a data.table with the columns:

AgeGroup

character. The age group the estimate is reported for. May be age or plus group

Age

integer. The lower age the estimate is reported for. May be an age or lower limit of plus group (inclusive)

<Statistic>

A reported statistic

SD

Standard deviation for the reported statistic.

Low

The lower limit of the estimated interval for the reported statistic.

High

The higher limit of the estimated interval for the reported statistic.

...

Any aggregation variables. The names of these are listed in 'GroupingVariables'

'GroupingVariables' is a data.table with a column containing the names of any aggregation variables.