Sampling Report data (ReportFdaSamplingData)

list with tow members:

GroupingVariables

a data.table with the variables used for aggregation in 'FishereisSampling' stored in the column 'GroupingVariables'

FisheriesSampling

a data.table described below.

FisheriesSampling is a report of sampling against total landings for partitions of a fishery. The report is a data.table with columns:

...

A column for each of the provided Aggregation variables

LandedRoundWeight

Total landings in kg

Catches

Number of catches sampled

Vessels

Number of vessels sampled

WeightMeasurments

Number of fished measured for weight

LengthMeasurments

Number of fished measured for length

AgeReadings

Number of fished with age determined

WeightOfSampledCatches

Total weight of the sampled catches