Run Reca Models

Runs prediction (catch-at-age estimate) for parameterized Reca models.

RunRecaModels(
  RecaParameterData,
  StoxLandingData,
  AggregationVariables = character(),
  TemporalResolution = c("Quarter", "Month"),
  Caa.burnin = numeric(),
  Seed = numeric()
)

Arguments

RecaParameterData
Parameters for Reca models.

StoxLandingData
Landings data (StoxLandingData).

AggregationVariables
character vector identifying columns in 'StoxLandingData' that results should be provided for.

TemporalResolution
default "Quarter", code for temporal resolution in landings: "Month" or "Quarter". Regulates temporal resolution for calculating fractional ages of fish. Not to be confused with any temporal covariate.

Caa.burnin
see documentation for eca.predict. Defaults to 0.

Seed
see documentation for eca.estimate. Defaults to seed stored in 'RecaParameterData'.

Value

RecaCatchAtAge

Details

Parameters may be obtained with ParameterizeRecaModels. If the function-paramter 'AggregationVariables' is provided, predictions will be provided for corresponding partitions of landings. The parameter 'StoxLandingData' may differ from the landings used in parameterisation (passed to ParameterizeRecaModels), as long as all not additional values / levels for the model covariates / effects are introduced.

If The models are configured for stock-splitting analysis. The variable 'Stock' will be added to 'AggregationVariables' in the return value (RecaCatchAtAge)

See also