Stock splitting parameters (StockSplittingParameters)

Table (data.table)

Defining parameters for the stock-splitting analysis in Reca, including parameters for the probability of misclassifying when determining stock membership of a specimen.

The stock splitting analysis allows catch at age to be estimated for two domains that partition all individuals, and that are observed for all age-determined specimens. It was developed for disciminating Coastal Cod and North East Arctic Cod, based on otholith growth patterns, and naming conventions are derived from that. It could be adapted to other stocks and in principle to other bipartite domain definitions (such as Sex).

Two otolith patterns are defined for each of the two stocks 'CC' and 'S'. Otolith type 1 and 2 identifies that a specimen belongs to the stock 'CC', and are interpreted by otoloith readers as 'certain' and 'uncertain' CC, respectively. Otolith type 4 and 5 identifies that a specimen belongs to the stock 'S', and are interpreted as 'uncertain' and 'certain' S, respectively.

StockNameCC

Name of the stock identified as CC

StockNameS

Name of the stock identified as S

ProbabilityType1As1

Probability of classifying a type 1 specimen as type 1 (certain CC).

ProbabilityType5As1

Probability of classifying a type 5 (certain S) specimen as type 1 (certain CC).

ProbabilityType2As2

Probability of classifying a type 2 (uncertain CC) specimen as type 2 (uncertain CC).

ProbabilityType4As2

Probability of classifying a type 4 (uncertain S) specimen as type 2 (uncertain CC).

ProbabilityType2As4

Probability of classifying a type 2 (uncertain CC) specimen as type 4 (uncertain S).

ProbabilityType4As4

Probability of classifying a type 4 (uncertain S) specimen as type 4 (uncertain S).

ProbabilityType1As5

Probability of classifying a type 1 (certain CC) specimen as type 5 (certain S).

ProbabilityType5As5

Probability of classifying a type 5 (certain S) specimen as type 5 (certain S).

The probabilities for different ways to classify a type must sum to 1. E.g.: ProbabilityType1As1 + ProbabilityType1As5 = 1.

The data table contains only one row.