Data from Norwegian port sampling program.

Norwegian port sampling for Haddock landings in 2018 of fresh whitefish north of Lat 64, covering Jig, Gillnet, Longline, and Demershal Seining. Sampling is implemented by a research vessel systematically sampling ports accepting fresh fish along the northern Norwegian coast. Landings are almost exclusively day-catches located to small areas. Vessels are intercepted during landing, and samples of up to 30 fish are taken stratified by gear and species. Length, weight and age measured for all fish. Sex and maturity when presentation allows for it. Weight of individual fish is reported as measured (presentation encoded in SApres), while total weight of landing and sample is converted to Whole-weight using constant factors standardized by the Norwegian Directorate of Fisheries (See table 2.1 in Conversion factors. Document is mostly in Norwegian, but the relevant table is in English as well).

data(catchsamples)

Format

Column names defined as in RDBES data model version 1.17 (RDBES data model v 1.17)

Examples

data(catchsamples)
table(catchsamples$SApres)

#> 
#> Gutted  Whole 
#>    714   1336