Catch at age traces

Compiles traces for age groups. That is the value for total catch at age in each iteration of eca.predict

makeAgeTracesRECA(prediction, unit = "millions", plusGroup = NULL)

Arguments

prediction
as returned by eca.predict or runRECA.

unit
unit of traced estimates. See details.

plusGroup
Fish this age or older will be grouped in one trace.

Value

data.table() with one column for each age (or plus group), and one row for each iteration of eca.predict.

Details

The number of iterations of eca.predict is determined by the parameter 'nSamples' to eca.estimate or runRECA and the parameter 'caa.burnin' (different from parameter 'burnin') to eca.predict or runRECA

parameter 'unit' supports:

number

Catch at age as number of fish

thousands

Catch at age as number of fish in thousands

millions

Catch at age as number of fish in millions

kg

Catch at age as mass in kilogrammes

T

Catch at age as mass in tons

kT

Catch at age as mass in kilotonnes

Examples

 data(recaPrediction)
 makeAgeTracesRECA(recaPrediction, plusGroup=13)

#>               2          3         4          5          6           7
#>  1: 0.006304755 0.05959851 0.1529576 0.05279517 0.01725642 0.006991260
#>  2: 0.004861709 0.07096931 0.1487804 0.04630795 0.01472454 0.008044174
#>  3: 0.004722424 0.06492861 0.1478442 0.04734141 0.01970048 0.005716015
#>  4: 0.006101567 0.06265084 0.1458517 0.04715619 0.01899657 0.006207489
#>  5: 0.006617257 0.06059345 0.1561691 0.04984283 0.02017890 0.006923694
#>  6: 0.003952131 0.06232255 0.1537817 0.05098235 0.02027646 0.006926204
#>  7: 0.005044708 0.06605677 0.1436042 0.05158375 0.01952983 0.007326283
#>  8: 0.004977607 0.04815270 0.1678995 0.04197466 0.01815273 0.007725440
#>  9: 0.004309134 0.05306774 0.1571335 0.04435640 0.01860007 0.009246991
#> 10: 0.004614078 0.06242579 0.1414158 0.05312626 0.02298217 0.006854556
#> 11: 0.004363612 0.06497029 0.1568845 0.04486985 0.01705601 0.008807803
#> 12: 0.004912125 0.06468540 0.1583453 0.04803690 0.01611460 0.006367487
#> 13: 0.006336998 0.06042861 0.1516472 0.04962195 0.01912797 0.006417346
#> 14: 0.005030575 0.07214481 0.1428219 0.04767875 0.02012531 0.006342532
#> 15: 0.004434211 0.06337705 0.1420016 0.04851749 0.02063353 0.005225620
#> 16: 0.003689138 0.06541279 0.1543920 0.04888367 0.01711159 0.004525207
#> 17: 0.004424446 0.07142289 0.1381678 0.04831339 0.01923587 0.005489268
#> 18: 0.005510280 0.06274178 0.1463929 0.05082353 0.02057401 0.007581038
#> 19: 0.005791505 0.06045619 0.1517977 0.04304034 0.02019732 0.007639292
#> 20: 0.005661578 0.05399410 0.1558372 0.04687118 0.02220149 0.008559748
#> 21: 0.005662560 0.05651505 0.1481751 0.04618082 0.01594151 0.010255065
#> 22: 0.005578002 0.06631573 0.1493245 0.04239868 0.01770724 0.007152166
#> 23: 0.005573501 0.07182893 0.1459966 0.04837092 0.02174691 0.006599028
#> 24: 0.004430997 0.06317910 0.1513009 0.04872974 0.01978211 0.005463782
#> 25: 0.004381467 0.06306983 0.1556676 0.04337875 0.01827546 0.005565660
#> 26: 0.003585581 0.06046277 0.1541168 0.04997727 0.01749955 0.007256625
#> 27: 0.005645521 0.06532214 0.1342160 0.05739014 0.02107350 0.008854187
#> 28: 0.004471162 0.06411094 0.1446586 0.05115578 0.02109960 0.008309532
#> 29: 0.004695431 0.07379624 0.1439439 0.04878108 0.01783841 0.007679370
#> 30: 0.005591926 0.06105539 0.1464299 0.04708730 0.02348174 0.006917543
#> 31: 0.004855459 0.05903094 0.1564723 0.05056841 0.02055882 0.006521097
#> 32: 0.004321371 0.06089690 0.1528216 0.04745930 0.01723595 0.009039477
#> 33: 0.005973288 0.06535908 0.1404395 0.04715282 0.02166820 0.010271852
#> 34: 0.005083896 0.06659423 0.1543714 0.04354137 0.01953815 0.006675013
#> 35: 0.005581340 0.06440214 0.1441122 0.05312360 0.01913420 0.008785514
#> 36: 0.005744314 0.05605868 0.1474881 0.04802619 0.02107628 0.008529596
#> 37: 0.005162903 0.06191700 0.1665753 0.04095757 0.01832711 0.006667357
#> 38: 0.005808516 0.06431134 0.1540857 0.04447239 0.01784421 0.008006472
#> 39: 0.006841286 0.05595774 0.1539924 0.05486535 0.01598310 0.007627637
#> 40: 0.005821281 0.06702787 0.1372795 0.05728853 0.01993019 0.006206140
#> 41: 0.004925012 0.06771525 0.1386242 0.04796351 0.01972909 0.008022402
#> 42: 0.005596717 0.06831115 0.1371785 0.05004992 0.02069630 0.007249139
#> 43: 0.003386395 0.06171583 0.1620446 0.03989145 0.01482555 0.007951782
#> 44: 0.003722741 0.06439814 0.1512275 0.04488535 0.01699218 0.008667787
#> 45: 0.005946050 0.06796418 0.1486430 0.05292436 0.01444666 0.007438895
#> 46: 0.004898165 0.06227881 0.1596461 0.04500683 0.01945253 0.004678176
#> 47: 0.004108075 0.05748746 0.1462109 0.05544567 0.01989314 0.006772089
#> 48: 0.003402328 0.06588649 0.1411686 0.05098767 0.01690380 0.009009475
#> 49: 0.005491198 0.06584059 0.1499654 0.04916274 0.01652099 0.007610129
#> 50: 0.004909195 0.06531986 0.1520562 0.04882820 0.01855231 0.005891370
#>               2          3         4          5          6           7
#>               8            9 10           11           12          13+
#>  1: 0.002150629 0.0011525931  0 1.217771e-04 6.144648e-04 0.000000e+00
#>  2: 0.002358353 0.0015597114  0 1.078304e-04 6.098980e-04 9.410669e-09
#>  3: 0.002440062 0.0007496458  0 6.383623e-05 2.882072e-04 0.000000e+00
#>  4: 0.002773213 0.0009163183  0 9.430809e-05 2.705135e-04 7.441141e-08
#>  5: 0.002155150 0.0016547566  0 8.614187e-05 2.620539e-04 0.000000e+00
#>  6: 0.001843145 0.0014203752  0 1.122648e-04 2.564113e-04 0.000000e+00
#>  7: 0.002444653 0.0013531923  0 3.426090e-04 2.282270e-04 0.000000e+00
#>  8: 0.002793213 0.0008860442  0 1.411574e-04 3.267035e-04 0.000000e+00
#>  9: 0.002040257 0.0010699635  0 1.265621e-04 3.177630e-04 0.000000e+00
#> 10: 0.002258373 0.0013307959  0 1.223007e-04 2.130684e-04 2.609496e-06
#> 11: 0.002303661 0.0008025198  0 7.358909e-05 1.749055e-04 1.349252e-06
#> 12: 0.003160813 0.0009096903  0 1.362778e-05 5.559235e-04 7.975127e-07
#> 13: 0.002228812 0.0012456823  0 2.591994e-05 3.706264e-04 2.295113e-06
#> 14: 0.002233462 0.0014523389  0 5.955834e-05 4.589794e-04 9.995179e-06
#> 15: 0.002963986 0.0016561667  0 8.581463e-05 1.184769e-04 5.081396e-06
#> 16: 0.003122307 0.0015961166  0 3.628166e-05 1.538328e-04 2.061823e-06
#> 17: 0.004623607 0.0014903125  0 6.521288e-05 3.134413e-04 1.270174e-05
#> 18: 0.003570027 0.0016997679  0 3.911207e-05 3.914723e-04 6.003296e-05
#> 19: 0.002088487 0.0012530571  0 5.343846e-06 3.519957e-04 1.259138e-05
#> 20: 0.001840244 0.0012026241  0 4.692764e-06 3.175222e-04 3.451735e-05
#> 21: 0.001953232 0.0014011609  0 3.915709e-05 3.036542e-04 7.011560e-05
#> 22: 0.001750681 0.0012623509  0 4.064611e-05 2.548211e-04 1.010794e-04
#> 23: 0.001955841 0.0013598742  0 7.310797e-06 3.086237e-05 4.371478e-05
#> 24: 0.002145601 0.0009634383  0 1.445929e-05 1.000568e-05 2.653620e-05
#> 25: 0.002678003 0.0005863868  0 5.165162e-05 3.282372e-05 4.743807e-05
#> 26: 0.002680228 0.0009156484  0 2.943688e-05 5.617073e-05 2.172267e-05
#> 27: 0.002393937 0.0006741418  0 6.185416e-05 1.931914e-04 3.486925e-06
#> 28: 0.002931751 0.0007718886  0 4.339301e-05 1.155875e-04 5.291413e-06
#> 29: 0.003236916 0.0006652107  0 3.490830e-05 6.371953e-05 3.471735e-06
#> 30: 0.003904778 0.0011062861  0 2.631163e-05 2.019382e-04 3.589604e-05
#> 31: 0.002418195 0.0010765559  0 1.378614e-05 8.728836e-05 9.913321e-05
#> 32: 0.002307902 0.0015375036  0 4.626279e-05 1.086063e-04 4.080373e-05
#> 33: 0.002657706 0.0013540534  0 7.135229e-05 1.928890e-04 1.138680e-05
#> 34: 0.002634866 0.0013468260  0 8.553653e-06 1.041812e-04 5.751264e-06
#> 35: 0.002364467 0.0017002701  0 9.946225e-06 6.736118e-05 7.550185e-06
#> 36: 0.002652070 0.0017941822  0 3.273570e-05 5.145349e-05 5.017649e-06
#> 37: 0.002600793 0.0014471137  0 2.140923e-05 1.174663e-05 1.929653e-06
#> 38: 0.002097910 0.0009733666  0 1.078697e-05 2.255779e-05 3.968827e-06
#> 39: 0.002209779 0.0007950670  0 1.457465e-05 2.534089e-06 7.473310e-07
#> 40: 0.002274936 0.0006747584  0 7.545239e-06 1.815289e-07 6.532469e-08
#> 41: 0.002283094 0.0007801623  0 1.386654e-05 4.022925e-06 4.971072e-08
#> 42: 0.002646830 0.0005936542  0 3.194617e-05 4.174879e-06 9.674614e-07
#> 43: 0.003427964 0.0007620603  0 3.694619e-05 2.558179e-06 4.451035e-06
#> 44: 0.003520539 0.0009178330  0 4.625905e-05 6.637687e-06 3.916492e-06
#> 45: 0.003681463 0.0014410094  0 8.703202e-05 9.129057e-06 9.667551e-07
#> 46: 0.003570202 0.0018540612  0 2.318478e-05 8.459398e-06 3.647615e-07
#> 47: 0.002877346 0.0014572615  0 1.631312e-05 2.109117e-06 1.579465e-07
#> 48: 0.002317154 0.0018601628  0 3.624391e-05 2.353083e-06 4.804117e-06
#> 49: 0.002993015 0.0027229258  0 5.954910e-05 1.707485e-05 7.549406e-06
#> 50: 0.002595426 0.0021873632  0 2.008719e-05 1.543113e-07 6.308757e-07
#>               8            9 10           11           12          13+