Merges polygons

Merge SpatialPolygonsDataFrame, such as StratumPolygon

mergePolygons(shape, mergeCol)

Arguments

shape
SpatialPolygonsDataFrame stratadefinition to convert

mergeCol
name of column that should be used for merging, all polygons with the same value in this column will be merged into one.

Value

SpatialPolygonsDataFrame with polygons merged

Details

All columns must have the same value for all polygons that are to be merged. If columns are not consistent in this regard, an error is raised.