Parameterize Reca.

Runs estimation of parameters for Reca models. Invokes eca.estimate.

ParameterizeRecaModels(
  RecaData,
  Nsamples = integer(),
  Burnin = integer(),
  Thin = integer(),
  ResultDirectory = character(),
  Lgamodel = c("log-linear", "non-linear"),
  Delta.age = numeric(),
  Seed = numeric()
)

Arguments

RecaData
RecaData as returned from PrepareRecaEstimate

Nsamples
number of MCMC samples that will be made available for eca.predict. See documentation for eca.estimate,

Burnin
number of MCMC samples run and discarded by eca.estimate before any samples are saved. See documentation for eca.estimate.

Thin
controls how many iterations are run between each samples saved. Defaults to 0. This may be set to account for autocorrelation introduced by Metropolis-Hastings simulation. see documentation for eca.estimate

ResultDirectory
a directory where Reca may store temp-files eca.estimate and eca.predict.

Lgamodel
The length age relationship to use for length-age fits (options: "log-linear", "non-linear": Schnute-Richards model). See documentation for eca.estimate.

Delta.age
see documentation for eca.estimate. Defaults to 0.001.

Seed
see documentation for eca.estimate. Defaults to random seed.

Value

RecaParameterData results from Reca Model Parameterization.

Details

eca.estimate performs Markov-chain Monte Carlo (MCMC) simulations to determine maximum likelihood of parameters for the given samples. This is computationally intensive and run time may be noticable. For a given model configuration running time is mainly determined by the parameters 'Nsample', 'Burnin' and 'Thin'.

temporary files

Various report functions may use output of this function with the function eca.predict which samples the posterior distributions of parameters. Communication between eca.estimate and eca.predict is managed by writing and reading files, and a directory for storing intermediate calculations must be provided with the parameter 'ResultDirectory'. For multi-chain analysis, a different directory should be provided for each chain.

See also