Prepare data for Reca.

Performs data checks and data conversions, and stores some data-related parameters in preparation for running ParameterizeRecaModels.

PrepareRecaEstimate(
  StoxBioticData,
  StoxLandingData,
  FixedEffects = character(),
  RandomEffects = character(),
  UseCarEffect = F,
  CarEffect = character(),
  CarNeighbours,
  UseAgingError = F,
  AgeErrorMatrix,
  UseStockSplitting = F,
  UseStockSplittingError = F,
  StockSplittingParameters,
  CellEffect = c("Off", "All"),
  MinAge = integer(),
  MaxAge = integer(),
  MaxLength = numeric(),
  LengthResolution = numeric(),
  HatchDay = integer()
)

Arguments

StoxBioticData
StoxBioticData data with samples from fisheries and approriate columns appended for specifying cells. See details.

StoxLandingData
StoxLandingData data with landings from fisheries and approriate columns appended for specifying cells. See details.

FixedEffects
optional, vector identifying column names that should be treated as fixed effects. See details

RandomEffects
optional, vector identifying column names that should be treated as random effects. See details

UseCarEffect
if TRUE, CAR-effect (conditional autoregressive effect) will be used.

CarEffect
optional, character identifying the column name that should be treated as CAR-effect. Mandatory if 'UseCarEffect' is TRUE.

CarNeighbours
CarNeighbours, mandatory if 'carEffect' is given. Identifies which values of the carEffect are to be considered as neighbours.

UseAgingError
If TRUE, aging, error parameters will be incorporated in the models.

AgeErrorMatrix
AgeErrorMatrix, optional, specifies the probabilities of misreading ages. mandatory if UseAgingError is TRUE.

UseStockSplitting
If TRUE, models will be condigured to provide estimates for each of two stocks based on otholitt-typing. See StockSplittingParameters.

UseStockSplittingError
If TRUE, the model for error in stock classification (otolithtypes) will be applied.

StockSplittingParameters
Parameters for stock splitting. Mandatory if 'UseStockSplitting' is TRUE. May be obtained with DefineStockSplittingParameters.

CellEffect
Configures the cell effect. If 'All', an interaction term will be added with all covariates that in the cell (whether they are fixed or random effects). Any CAR-effect is always included in the cell effect.

MinAge
optional, must match dimensions of any 'AgeErrorMatrix'. If not provided it will be derived from data.

MaxAge
optional, must match dimensions of any 'AgeErrorMatrix'. If not provided it will be derived from data.

MaxLength
optional, maximal fish length in data in cm. If not provided it will be derived from data.

LengthResolution
optional, resolution for length measurements in cm. If not provided modal value from data is used.

HatchDay
defaults to 1 representing Jan 1st. encoding the day of the year when fish is consider to transition from one age to the next.

Value

RecaData Data prepared for running Reca.

Details

Parameters are obtained for 'cells' using ParameterizeRecaModels and and applied to landings-data when predicting total catch-at-age using RunRecaModels.

'Cells' are defined by all effects / covariates that are specified for both 'StoxLandingData' and 'StoxBioticData' (see parameters 'FixedEffects', 'RandomEffects', and 'CarEffect'). That is, any variable that are to be included in the cell definition must always exist (not containing NAs) and be coded coherently in both 'StoxBioticData' and 'StoxLandingData'.

Only effects that are part of the cell definition may be specified as 'Fixed effect' or 'CAR effect'. For fixed effect all values / levels of the effect must be sampled. For CAR effect, cell or neighbouring cells ('CarNeighbours') must be sampled

Any variable in samples 'StoxBioticData' may be included as a random effect as long as it is always observed (not containing NAs).

A random effect identifying haul / landing is always included in Reca. Do not add haul-identifiers as covariates.

The option UseStockSplitting requires that a Variable 'otolithtype' is added to the table 'Individual' in StoxBioticData.

See also