Report convergence Statistics for parameter simulations

For multi-chain parameter simulations, intra-chain variance, inter-chain variance and the Gelman-Rubins R statistic is reported.

'chains' in this respect refers to statistically independent simulations.

ReportParameterConvergence(
  ParameterizationSummaryData,
  Tolerance = numeric(),
  Decimals = integer()
)

Arguments

ParameterizationSummaryData

summary statistics for Reca parameters

Tolerance

threshold for reporting parameters. Defaults to 0.1. See details

Decimals

integer specifying the number of decimals to report for 'GelmanRubinR'. Defaults to 2

Value

ParameterConvergenceData

Details

Only parameters with a Gelman-Rubins R statistic that deviates sufficiently from 1 is reported. The accepted deviation is controlled by the argument 'Tolerance'. Parameters are reported if they differ from 1 by 'Tolerance' or more, so that a Tolerance of 0 reports all parameters. If not provided a default Tolerance of 0.1 is used.

Gelman-Rubins R is described in Gelman and Rubin (Statistical Science, 1992): DOI: https://doi.org/10.1214/ss/1177011136

The Gelman-Rubins R reported here is approximately that specified in Gelman and Rubins Eq. 20, ignoring the contribution from the factor (df/(df-2)).

In the report InterVariance correspond to their B/n and IntraVariance correspond to their W.

See also