Report catch at length

Tabulates summary statistics for total catch (number) at length from MCMC simulations using Reca. MCMC simulations are typically obtained with RunRecaModels. Summary statistics are obtained from the posterior distribution, and the interval is reported as equal-tailed credible intervals.

Different length groups than the ones reported in the argument 'RecaCatchAtAge' may be specified with the argument 'IntervalWidth'. This will specify equi-intervalled lengthgroups with the smallest lengthgroup starting at 0. If it does not align with the length groups reported in 'RecaCatchAtAge' length group assignment is done to the highest overlapping length group.

If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as area, gear, stock, etc., summary statistics will be presented similarly.

Rounding of numbers according to the argument 'Decimals' is done with round, so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

The units considered valid for catch at length in numbers are those listed for quantity 'cardinaltiy' in StoxUnits

ReportRecaCatchAtLength(
  RecaCatchAtAge,
  IntervalWidth = numeric(),
  Decimals = integer(),
  Unit = RstoxData::getUnitOptions("cardinality", conversionRange = c(1, 1e+12)),
  LengthInterval = numeric()
)

Arguments

RecaCatchAtAge

Results from MCMC simulations (RecaCatchAtAge).

IntervalWidth

The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.

Decimals

integer specifying the number of decimals to report for 'CatchAtLength', 'SD', 'Low' and 'High'. Defaults to zero.

Unit

unit for 'CatchAtLength', 'SD', 'Low' and 'High'

LengthInterval

width of length bins in cm. If not provided, the interval in 'RecaCatchAtAge' will be used.

Value

ReportFdaCatchAtAgeData

See also