Report catch statistics

Report summary statistics for landed catches from MCMC simulations using Reca. Summary statistics are obtained from the posterior distribution, and the interval is reported as equal-tailed credible intervals are reported.

Weights are reported in kg, lengths in cm.

If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as area, gear, stock, etc., summary statistics will be presented similarly.

Rounding of numbers according to the argument 'Decimals' is done with round, so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

The units considered valid for mean lengths are those listed for quantity 'length' in StoxUnits The units considered valid for weights are those listed for quantity 'mass' in StoxUnits The units considered valid for total catch in numbers are those listed for quantity 'cardinality' in StoxUnits

ReportRecaCatchStatistics(
  RecaCatchAtAge,
  IntervalWidth = numeric(),
  DecimalOptions = FALSE,
  DecimalTotalNumber = integer(),
  DecimalTotalWeight = integer(),
  DecimalMeanAge = integer(),
  DecimalMeanWeight = integer(),
  DecimalMeanLength = integer(),
  UnitOptions = FALSE,
  UnitTotalNumber = RstoxData::getUnitOptions("cardinality", conversionRange = c(1,
    1e+12)),
  UnitTotalWeight = RstoxData::getUnitOptions("mass", conversionRange = c(1, 1e+12)),
  UnitMeanWeight = RstoxData::getUnitOptions("mass", conversionRange = c(1e-04, 10)),
  UnitMeanLength = RstoxData::getUnitOptions("length", conversionRange = c(1e-04, 10))
)

Arguments

RecaCatchAtAge

Results from MCMC simulations (RecaCatchAtAge).

IntervalWidth

The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.

DecimalOptions

logical determining whether the StoX user interface should show decimal options

DecimalTotalNumber

integer specifying the number of decimals to report for 'TotalNumber', and the corresponding 'SD', 'Low' and 'High'. to 0

DecimalTotalWeight

integer specifying the number of decimals to report for 'TotalWeightDefaults', and the corresponding 'SD', 'Low' and 'High'. to 0

DecimalMeanAge

integer specifying the number of decimals to report for 'MeanIndividualAge', and the corresponding 'SD', 'Low' and 'High'. Defaults to 1

DecimalMeanWeight

integer specifying the number of decimals to report for 'MeanIndividualWeight', and the corresponding 'SD', 'Low' and 'High'. Defaults to 3

DecimalMeanLength

integer specifying the number of decimals to report for 'MeanIndividualLength', and the corresponding 'SD', 'Low' and 'High'. Defaults to 2

UnitOptions

logical determining whether the StoX user interface should show unit options

UnitTotalNumber

unit for total catch in numbers. Defaults to Mi (millions)

UnitTotalWeight

unit for weight of total catch. Defaults to kt

UnitMeanWeight

unit for mean weight. Defaults to kg

UnitMeanLength

unit for mean length. Defaults to cm.

Value

ReportFdaSummaryData

See also