Report weight at age

Tabulates summary statistics for mean weights at age from MCMC simulations using Reca.

If 'RecaCatchAtAge' contains estimate for a set of aggregation variables, such as area, gear, stock, etc., summary statistics will be presented similarly.

Mean weight for plus-groups are a weighted by the relative catch-at-age in each composite age group. For iterations where all of the plus-group ages have a zero catch at age, this weight is not defined, and summary statistics are obtained from the remaining iterations.

Rounding of numbers according to the argument 'Decimals' is done with round, so that negative numbers specify rounding to powers of ten, and rounding of the digit 5 is towards the even digit.

The units considered valid for mean weights are those listed for quantity 'mass' in StoxUnits

MCMC simulations are typically obtained with RunRecaModels. Summary statistics are obtained from the posterior distribution, and the interval is reported as 90% equal-tailed credible intervals.

ReportRecaWeightAtAge(
  RecaCatchAtAge,
  PlusGroup = integer(),
  IntervalWidth = numeric(),
  Decimals = integer(),
  Threshold = numeric(),
  Unit = RstoxData::getUnitOptions("mass", conversionRange = c(1e-04, 10))
)

Arguments

RecaCatchAtAge

Results from MCMC simulations (RecaCatchAtAge).

PlusGroup

If given, ages 'PlusGroup' or older are included in a plus group.

IntervalWidth

The width of the reported credible interval. Defaults to 0.9 for 90 per cent credible intervals.

Decimals

integer specifying the number of decimals to report for 'MeanIndividualWeight', 'SD', 'Low' and 'High'. Defaults to 2.

Threshold

threshold for reporting mean weight. Rows with an estimated Catch At Age (number of individuals) lower than this will have NA reported for their mean weight. Defaults to 0.

Unit

unit for 'MeanIndividualWeight', 'SD', 'Low' and 'High'

Value

ReportFdaWeightAtAgeData

See also