Run Reca Models

Runs prediction (catch-at-age estimate) for parameterized Reca models.

RunRecaModels(
  RecaParameterData,
  StoxLandingData,
  GroupingVariables = character(),
  TemporalResolution = c("Quarter", "Month"),
  Caa.burnin = numeric(),
  Seed = numeric(),
  CollapseLength = TRUE
)

Arguments

RecaParameterData

Parameters for Reca models.

StoxLandingData

Landings data (StoxLandingData).

GroupingVariables

character vector identifying columns in 'StoxLandingData' that results should be provided for.

TemporalResolution

default "Quarter", code for temporal resolution in landings: "Month" or "Quarter". Regulates temporal resolution for calculating fractional ages of fish. Not to be confused with any temporal covariate.

Caa.burnin

see documentation for eca.predict. Defaults to 0.

Seed

see documentation for eca.estimate. Defaults to seed stored in 'RecaParameterData'.

CollapseLength

indicates whether length groups should be collapsed in result. Defaults to TRUE

Value

RecaCatchAtAge

Details

Parameters may be obtained with ParameterizeRecaModels. If the function-parameter 'GroupingVariables' is provided, predictions will be provided for corresponding partitions of landings. The parameter 'StoxLandingData' may differ from the landings used in parameterisation (passed to ParameterizeRecaModels), as long as all not additional values / levels for the model covariates / effects are introduced.

If The models are configured for stock-splitting analysis. The variable 'Stock' will be added to 'GroupingVariables' in the return value (RecaCatchAtAge)

If the 'GroupingVariables' specify a large number of partitions of the landings, this function may exhaust available computer memory.

By default length groups are collapsed into one length group in the result. The full age-length prediction may be extracted by setting the parameter 'CollapseLength' to FALSE. This will increase the risk of exhausting available computer memory.

See also