Catch at age traces

Compiles traces for age groups. That is the value for total catch at age in each iteration of eca.predict

makeAgeTracesRECA(prediction, unit = "millions", plusGroup = NULL)

Arguments

prediction

as returned by eca.predict or runRECA.

unit

unit of traced estimates. See details.

plusGroup

Fish this age or older will be grouped in one trace.

Value

data.table() with one column for each age (or plus group), and one row for each iteration of eca.predict.

Details

The number of iterations of eca.predict is determined by the parameter 'nSamples' to eca.estimate or runRECA and the parameter 'caa.burnin' (different from parameter 'burnin') to eca.predict or runRECA

parameter 'unit' supports:

number

Catch at age as number of fish

thousands

Catch at age as number of fish in thousands

millions

Catch at age as number of fish in millions

kg

Catch at age as mass in kilogrammes

T

Catch at age as mass in tons

kT

Catch at age as mass in kilotonnes

See also

Examples

 data(recaPrediction)
 makeAgeTracesRECA(recaPrediction, plusGroup=13)
#>            2        3        4        5        6        7         8         9
#>  1: 5.071436 57.52779 157.2190 49.13707 17.90393 6.670291 2.0559312 0.4991370
#>  2: 6.748423 55.24290 145.8560 46.99602 23.01269 7.917396 2.3648727 0.6610583
#>  3: 7.321326 65.41784 150.4459 42.91064 23.27061 6.567285 2.3985501 0.7675303
#>  4: 7.597182 56.39696 151.1578 55.97815 16.74334 8.541702 1.2456728 0.4381584
#>  5: 7.877592 72.62752 140.9677 46.70734 19.94819 7.020340 0.8478461 0.7655145
#>  6: 6.528875 62.79681 149.4854 46.99648 19.26803 6.740761 1.6851874 1.0920434
#>  7: 4.293750 73.08370 144.4609 46.18798 19.52897 6.541902 2.1690896 1.0238879
#>  8: 3.733232 62.05030 151.7276 46.84826 18.67632 7.624992 3.5723253 1.2046359
#>  9: 3.121939 62.78314 159.5012 40.39313 19.26004 6.768792 2.3022494 1.7607639
#> 10: 4.022617 68.41884 141.1321 51.47981 20.13147 8.184323 1.9370136 1.1923242
#> 11: 4.190463 62.59433 143.4548 52.33711 18.80929 7.358876 1.8556586 1.3782049
#> 12: 5.460523 70.86194 135.3595 52.57318 21.08608 6.950138 1.5939303 1.6721217
#> 13: 6.542439 57.90061 149.3361 44.33587 23.80260 6.560222 2.0832829 2.3353937
#> 14: 7.042423 59.59119 157.6821 48.25705 16.84770 6.796558 2.5172785 1.9308792
#> 15: 5.604691 63.46678 145.0272 46.74550 22.93742 5.315149 3.4076541 2.2946706
#> 16: 4.929835 57.39144 152.3403 43.74899 17.71521 6.780860 2.4607543 2.0223970
#> 17: 5.364131 51.70178 151.2191 50.61327 19.63723 6.239793 2.9911022 2.1954029
#> 18: 5.908145 65.56397 143.4936 46.84448 21.82124 7.024807 2.5761218 1.6359209
#> 19: 5.712961 63.32671 141.0058 53.12444 21.59879 6.504901 2.9211724 1.9044963
#> 20: 6.128067 65.93401 148.4191 44.32358 18.60884 6.149501 2.3475305 2.1413553
#> 21: 4.164608 64.18252 149.5307 52.50690 18.11855 5.007737 2.0195596 1.7734386
#> 22: 4.528860 75.05659 135.7944 51.88694 15.59514 6.122143 2.9722324 1.6671924
#> 23: 5.153530 59.16048 143.0562 54.39939 20.42541 5.172893 3.2385877 1.2980825
#> 24: 5.369316 65.44632 137.9879 50.33092 18.07390 7.626326 3.0974696 1.1416621
#> 25: 7.395269 65.37642 146.6788 49.70483 16.74613 6.258194 2.3330227 0.7026741
#> 26: 6.975377 62.48815 147.9744 45.33445 20.52193 5.387252 2.1436035 1.3995638
#> 27: 4.729653 66.28890 143.0541 48.25146 19.69965 6.503520 2.8543996 1.9872122
#> 28: 5.630833 48.87601 155.3567 49.21233 21.10252 7.047630 4.7795140 1.2666304
#> 29: 5.877816 61.46920 150.7409 50.05002 18.46491 6.397042 3.2287542 1.5830577
#> 30: 4.939111 58.38904 153.0760 47.86543 18.66318 7.617063 2.8305008 1.7062701
#> 31: 6.858497 75.00919 139.7398 45.91215 17.20359 9.373313 2.8127556 1.6350333
#> 32: 4.841475 68.75061 134.2295 55.80227 18.37724 8.860029 2.5046351 0.8445977
#> 33: 5.524771 65.32981 144.1201 52.92920 18.89557 7.049977 2.4429431 1.0579736
#> 34: 5.704655 65.17603 159.2678 43.89994 16.83023 7.768336 2.8471866 1.0555871
#> 35: 5.494653 61.89987 141.6850 47.24739 20.25099 7.527318 2.1813132 1.7732306
#> 36: 4.938364 63.68857 140.3199 46.96721 21.77508 5.430538 2.4653339 2.4858685
#> 37: 4.934527 60.08507 155.4755 46.21513 21.22857 4.621337 2.1725778 2.1679303
#> 38: 3.297387 73.50411 150.3488 44.24361 17.05974 4.643883 1.7255429 1.8185118
#> 39: 3.844261 64.04685 154.7931 45.88875 18.73706 8.191418 2.1978583 2.0499367
#> 40: 3.948560 67.06793 150.8472 43.97370 16.05150 8.411872 2.1358632 1.5869472
#> 41: 4.618407 58.66643 135.9670 57.55610 22.26289 7.085175 1.9030166 1.2970883
#> 42: 3.861240 61.37838 145.3992 50.93027 19.24546 6.708574 1.9868694 1.9775288
#> 43: 3.602793 55.67180 147.9914 52.23408 22.49276 8.391988 1.3312130 1.1922926
#> 44: 3.649253 65.10792 145.2303 40.92316 21.69071 9.624586 1.9222380 2.4141120
#> 45: 4.482321 52.17636 155.8123 49.97792 19.52461 7.288277 1.2714825 2.1021133
#> 46: 4.913193 52.93883 156.7286 47.38942 19.64794 5.301782 1.6810399 2.0681667
#> 47: 4.784660 62.48634 148.4956 45.87648 18.68233 8.034028 2.2043470 1.5077839
#> 48: 4.864327 67.13014 135.9789 52.34357 23.29429 6.909578 2.2946289 2.1153921
#> 49: 4.915051 54.00304 143.6959 60.94825 20.65901 6.215670 1.9823930 2.6315814
#> 50: 4.827608 68.98072 144.4407 48.40054 16.77356 7.143069 2.0929583 2.1556835
#>            2        3        4        5        6        7         8         9
#>     10           11           12          13+
#>  1:  0 5.358476e-01 5.235632e-02 2.574511e-03
#>  2:  0 3.746864e-01 9.829235e-02 1.577620e-03
#>  3:  0 3.970121e-01 7.230164e-02 3.041278e-04
#>  4:  0 3.993252e-01 2.537726e-02 3.619289e-03
#>  5:  0 5.322943e-01 4.107091e-03 1.387163e-02
#>  6:  0 8.467851e-01 3.907370e-04 5.673688e-03
#>  7:  0 3.222895e-01 1.420120e-05 4.119028e-03
#>  8:  0 1.232903e-01 0.000000e+00 3.595313e-04
#>  9:  0 1.376769e-01 3.613194e-05 4.156555e-03
#> 10:  0 5.201187e-02 0.000000e+00 3.110813e-03
#> 11:  0 8.012253e-02 0.000000e+00 4.303536e-04
#> 12:  0 1.841865e-01 1.304262e-03 1.203585e-02
#> 13:  0 9.143089e-02 8.009456e-03 6.342044e-02
#> 14:  0 1.206305e-01 8.072043e-02 8.818020e-02
#> 15:  0 1.034273e-01 6.230199e-02 4.979702e-02
#> 16:  0 1.238170e-01 6.378653e-02 3.979492e-02
#> 17:  0 2.485960e-01 7.471258e-02 5.202473e-02
#> 18:  0 1.721069e-01 7.526831e-02 2.152834e-02
#> 19:  0 3.579559e-01 2.819907e-02 1.773465e-02
#> 20:  0 2.825361e-01 9.746262e-03 3.592062e-03
#> 21:  0 2.734809e-01 1.900538e-02 3.050258e-03
#> 22:  0 2.849947e-01 9.705439e-02 8.726843e-03
#> 23:  0 1.288097e-01 4.751068e-02 8.352989e-04
#> 24:  0 5.559850e-02 2.273069e-02 5.046907e-03
#> 25:  0 1.828175e-02 2.125931e-02 1.447473e-02
#> 26:  0 5.337439e-03 4.240766e-02 2.362282e-02
#> 27:  0 2.163786e-05 8.580752e-03 8.331207e-02
#> 28:  0 0.000000e+00 2.049944e-02 5.919618e-02
#> 29:  0 0.000000e+00 1.101308e-01 7.039411e-02
#> 30:  0 3.419798e-05 1.299411e-01 4.425388e-02
#> 31:  0 1.852677e-03 2.694129e-02 8.776073e-02
#> 32:  0 3.271555e-03 1.795180e-02 3.418712e-02
#> 33:  0 7.359879e-03 4.021518e-02 7.920021e-02
#> 34:  0 9.226291e-03 5.134233e-02 3.729445e-02
#> 35:  0 4.137360e-02 2.386780e-02 5.442972e-02
#> 36:  0 3.578867e-02 1.843442e-02 1.208877e-01
#> 37:  0 1.619796e-02 2.765894e-03 4.369335e-02
#> 38:  0 6.628398e-03 1.245390e-03 8.743570e-03
#> 39:  0 2.932724e-03 2.074712e-02 5.107857e-03
#> 40:  0 5.133491e-03 2.923605e-02 7.094264e-05
#> 41:  0 1.064962e-01 3.131209e-02 2.814087e-03
#> 42:  0 2.243847e-01 1.505943e-02 2.487310e-03
#> 43:  0 2.094689e-01 3.107703e-02 7.739626e-05
#> 44:  0 2.206065e-01 8.049682e-02 2.361194e-03
#> 45:  0 7.956321e-02 3.518543e-02 2.606357e-03
#> 46:  0 3.598420e-02 2.699982e-02 4.732701e-04
#> 47:  0 1.766134e-02 1.948988e-02 3.597995e-05
#> 48:  0 2.554516e-02 2.844892e-02 9.588896e-04
#> 49:  0 7.162600e-02 5.526441e-02 2.982786e-05
#> 50:  0 6.614665e-02 4.390141e-02 4.158394e-04
#>     10           11           12          13+