Catch at Age result table

Produces common reported statistics for Catch-At-Age estimation.

makeResultTableRECA(
  prediction,
  unit = "millions",
  plusGroup = NULL,
  alpha = 0.05
)

Arguments

prediction
as returned by eca.predict or runRECA.

unit
unit of reported estimates. See details.

plusGroup
Fish this age or older will be grouped in report

alpha
value for percentiles.

Value

data.table() with columns:

age

Age statistics are reported for

total

Estimated total catch of age group

unit

Unit for total catch and standard deviation for estimated total

sd

Standard deviation for estimated total

cv

Coefficient of variation for estimated total

lowerQuantile

Lower quantile (100*alpha/2 percentile) for estimated total

UpperQuantile

Upper quantile: (100*(1-alpha/2) percentile) for estimated total

alpha

alpha for quantiles.

Details

parameter 'unit' supports:

number

Catch at age as number of fish

thousands

Catch at age as number of fish in thousands

millions

Catch at age as number of fish in millions

kg

Catch at age as mass in kilogrammes

T

Catch at age as mass in tons

kT

Catch at age as mass in kilotonnes

Examples

 data(recaPrediction)
 makeResultTableRECA(recaPrediction, plusGroup=13)
#>     age        total     unit upperQuantile lowerQuantile alpha
#>  1:   2 5.057110e-03 millions  6.554199e-03  3.443560e-03  0.05
#>  2:   3 6.329060e-02 millions  7.207373e-02  5.327617e-02  0.05
#>  3:   4 1.496392e-01 millions  1.655559e-01  1.372012e-01  0.05
#>  4:   5 4.836352e-02 millions  5.687389e-02  4.118642e-02  0.05
#>  5:   6 1.893061e-02 millions  2.280652e-02  1.474727e-02  0.05
#>  6:   7 7.291836e-03 millions  1.002825e-02  4.801351e-03  0.05
#>  7:   8 2.623102e-03 millions  3.854532e-03  1.840897e-03  0.05
#>  8:   9 1.247685e-03 millions  2.113743e-03  6.097544e-04  0.05
#>  9:  10 0.000000e+00 millions  0.000000e+00  0.000000e+00  0.05
#> 10:  11 5.370901e-05 millions  1.378734e-04  5.786410e-06  0.05
#> 11:  12 1.712547e-04 millions  5.977538e-04  6.152361e-07  0.05
#> 12: 13+ 1.406165e-05 millions  9.260425e-05  0.000000e+00  0.05